dash_app/pyproject.toml
[tool.poetry]
name = "dash-app"
version = "0.1.0"
description = ""
authors = ["Dominik Irimi <dominik.irimi@protonmail.ch>"]
readme = "README.md"
packages = [{include = "dash_app"}]

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.7.0"
pandas = "^1.5.1"
influxdb-client = "^1.34.0"
plotly = "^5.11.0"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
dash-auth = "^1.4.1"
arrow = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
